refactor(favorites): Move favorites-related code into a favorites package.
Leave the legacy FavoritesLiveData where it is.
