+++ app/src/main/kotlin/ca/rmen/android/poetassistant/main/dictionaries/rt/FavoritesLiveData.kt
@@ -22,7 +22,7 @@ package ca.rmen.android.poetassistant.main.dictionaries.rt
 import android.content.Context
 import android.util.Log
 import ca.rmen.android.poetassistant.Constants
-import ca.rmen.android.poetassistant.FavoritesRepository
+import ca.rmen.android.poetassistant.main.favorites.FavoritesRepository
 import ca.rmen.android.poetassistant.R
 import ca.rmen.android.poetassistant.di.NonAndroidEntryPoint
 import ca.rmen.android.poetassistant.main.dictionaries.ResultListData
