+++ app/src/main/kotlin/ca/rmen/android/poetassistant/main/favorites/Favorite.kt
@@ -17,7 +17,7 @@
  * along with Poet Assistant.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ca.rmen.android.poetassistant
+package ca.rmen.android.poetassistant.main.favorites
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
