+++ app/src/main/kotlin/ca/rmen/android/poetassistant/main/favorites/FavoritesRepository.kt
@@ -17,12 +17,13 @@
  * along with Poet Assistant.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ca.rmen.android.poetassistant
+package ca.rmen.android.poetassistant.main.favorites
 
 import android.content.Context
 import android.net.Uri
 import android.text.TextUtils
 import android.util.Log
+import ca.rmen.android.poetassistant.Constants
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
